lib file structure
