+++ config/initializers/ext.rb
@@ -1,5 +1,3 @@
 # frozen_string_literal: true
 
-Dir.glob(Rails.root.join('lib/ext/**/*.rb')).sort.each do |filename|
-  require filename
-end
+require 'gem_ext'
