+++ lib/gem_ext.rb
@@ -1,0 +1,1 @@
+require 'gem_ext/wondeclient'
